refactor: edit profile
