+++ frontend/src/pages/StudentDetails/components/EditProfile/EditProfile.tab.tsx
@@ -1,5 +1,6 @@
 import { FormEvent, useState } from 'react';
 
+import { Divider } from '@mui/material';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import {
@@ -8,11 +9,15 @@ import {
   EditAccountOptionsDTO,
 } from '#modules/account/api/editAccount';
 import { useCurrentUserData } from '#modules/student/hooks/useCurrentUser.data';
+import { Student } from '#modules/student/student.types';
+import { Avatar } from '#shared/components/Avatar/Avatar';
+import { FlexRow } from '#shared/components/FlexBox/FlexBox';
 import { LoadingButton } from '#shared/components/LoadingButton/LoadingButton';
 import { useTranslation } from '#shared/i18n/useTranslation';
 import { ApiFormError } from '#shared/infra/errors';
 
 import { EditProfileFormFields } from '../FormFields/EditProfileFormFields';
+import { StudentDetailsInfo } from '../StudentDetailsInfo';
 
 export function EditProfileTab() {
   const student = useCurrentUserData();
@@ -54,8 +59,44 @@ export function EditProfileTab() {
     mutate(formValues);
   }
 
+  function convertPictureToURL() {
+    if (formValues.picture) {
+      if (formValues.picture.name === '') {
+        return '';
+      } else {
+        return URL.createObjectURL(formValues.picture);
+      }
+    }
+    return student.picture;
+  }
+  function convertToPreview(form: EditAccountOptions): Student {
+    return {
+      description: form.description,
+      faculty: form.faculty,
+      name: form.firstName + ' ' + form.lastName.toUpperCase(),
+      id: -1,
+      path: form.path,
+      picture: convertPictureToURL(),
+      promo: form.promo,
+      socialLinks: [],
+      staff: false,
+      url: '',
+      username: form.username,
+    };
+  }
+
   return (
     <>
+      <FlexRow>
+        <Avatar
+          alt={formValues.firstName + ' ' + formValues.lastName.toUpperCase()}
+          src={convertPictureToURL()}
+          size="xl"
+          sx={{ m: 1 }}
+        />
+        <StudentDetailsInfo student={convertToPreview(formValues)} />
+      </FlexRow>
+      <Divider sx={{ my: 2 }} />
       <form id="edit-account-form" onSubmit={(e) => onSubmit(e)}>
         <EditProfileFormFields
           formValues={formValues}
